修复大量警告,用静态分析修复内存问题

+++ WeiBo/Classes/Compose(发微博)/View/CJEmotionTabBar.m
@@ -56,7 +56,7 @@
     
     
     if ([self.delegate respondsToSelector:@selector(emotionTabBar:didSelectedButton:)]) {
-        [self.delegate emotionTabBar:self didSelectedButton:button.tag];
+        [self.delegate emotionTabBar:self didSelectedButton:(int)button.tag];
     }
 }
